RE-279 Enable saving an instance as an image
In order to allow the capability to save single
use slaves as an image for re-use later, a job
added which does so.

A custom python script is used due to the fact
that pyrax appears to be the only reliable way
to do the action in RAX PubCloud. The nova CLI
does not seem to work and we need to use pyrax
for the rax Ansible modules. This can be revisited
when we switch to using the shade-based modules
instead.